UCM/UTIL/RELOC: Don't rely on RTLD_NEXT for lookup
Rather than relying on RTLD_NEXT to return the symbol and falling back
to RTLD_DEFAULT instead lookup the symbol from the current shared
library directly.

This resolves an issue where on RISC-V in some circumstances the symbol
returned by RTLD_NEXT is null and RTLD_DEFAULT then provides a pointer
into the PLT stub. Since the PLT stub is too small to be patched by
directly resolving the symbol here the PLT is avoided and the C library
function can be directly patched.

Signed-off-by: Rob Bradford <rbradford@rivosinc.com>