UCM: Fix RELOC patching method for shared libraries
Using UCX_MEM_MMAP_HOOK_MODE=reloc was crashing when running on RISC-V
because ucm_dl_populate_symbols() was reaching through a symbol table
using exactly the value it found out of the ELF dynamic section entry.
This entry is a virtual addres if and only if the image in question has
been loaded at address 0 (which is architecture dependent and not the
case on RISC-V). Treating it as as virtual address when not loaded at 0
is therefore not appropriate (the pointer is dangling.)

Add the dlpi_addr to dynamic entries that are treated as virtual
addresses to compute their correct value.

Signed-off-by: Evan Green <evan@rivosinc.com>
Co-developed-by: Rob Bradford <rbradford@rivosinc.com>