UCM/BISTRO: Use appropriate atomic swap for size of update
On RISC-V for an atomic update the destination must be naturally aligned
with the size of the type. Since the destinations in this case are
addresses of library functions to patch we cannot control the addresses.
Fortunately the instructions must be aligned to a 16-bit alignment which
also matches the size of the patch on RISC-V (compressed J instruction.)

For other architectures the size of patch is either 4 bytes (aarch64) or
2 bytes (x86-64) which are both handled by this patch.

Signed-off-by: Rob Bradford <rbradford@rivosinc.com>