UCM/UTIL/RELOC: Turn PT_DYNAMIC table entries to pointers on RISC-V
On some combinations of C library implementation and architecture the
entries in the PT_DYNAMIC table are not relocated and stored as
pointers. Instead the values are offsets from the address that the
library is loaded at. RISC-V with glibc is one such example.

Unfortunately since the internal glibc headers that expose whether the
value is relocated on not are not available the best solution is to make
this architecturally conditional. As this behaviour is now established
as part of the ABI it will not change in the future.

Since this should only be done for the values from the table that are
treated like pointers create a new function ucm_reloc_get_pointer() that
applies the architecture check was added

This issue was discovered when using UCX_MEM_MMAP_HOOK_MODE=reloc on RISC-V.

Fixes: #9283

Co-developed-by: Evan Green <evan@rivosinc.com>
Signed-off-by: Rob Bradford <rbradford@rivosinc.com>